requirements.txt
aiodns==3.2.0
aiofiles==0.8.0
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.17.0
alabaster==1.0.0
art==6.1
asgiref==3.8.1
asttokens==2.4.1
async-google-trans-new==1.0.3
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
asyncpraw==7.7.1
asyncprawcore==2.4.0
attr==0.3.2
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
barcode-generator==0.1rc15
beautifulsoup4==4.12.3
bidict==0.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
configparser==7.1.0
cryptography==43.0.3
decorator==5.1.1
discord.py==2.4.0
discord_localization==1.1.4
Django==5.1.3
django-recaptcha==4.0.0
docutils==0.21.2
executing==2.1.0
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
HTMLParser==0.0.2
idna==3.10
imagesize==1.4.1
ipython==8.12.3
ipywidgets==8.1.5
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.4
jupyterlab_widgets==3.0.13
keyring==25.2.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
multidict==6.1.0
numpy==2.1.3
packaging==24.2
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==5.9.8
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
pypokedex==1.6.0
pytemperature==1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.1
pywin32-ctypes==0.2.3
pyyaml
qrcode==7.4.2
redis==5.2.0
requests==2.32.3
simple-websocket==1.1.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
stack-data==0.6.3
thread==2.0.5
topggpy==1.4.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.2.3
urllib3-secure-extra==0.1.0
wavelink==3.4.1
wcwidth==0.2.13
widgetsnbextension==4.0.13
wsproto==1.2.0
yarl==1.17.1
